// ui-react/apps/console/src/client/types.gen.ts
export type NamespaceTenantId = string;
// ...
export type DeviceUid = string;

/**
 * Device's status
 */
export type DeviceStatus = 'accepted' | 'rejected' | 'pending' | 'removed' | 'unused';

/**
 * Device's identity
 */
export type DeviceIdentity = {
    /**
     * Device's MAC
     */
    mac: string;
};

/**
 * Device's info
 */
export type DeviceInfo = {
    /**
     * Device's OS name
     */
    id: string;
    /**
     * Device's OS pretty name
     */
    pretty_name: string;
    /**
     * Device's OS version
     */
    version: string;
    /**
     * Device's OS arch
     */
    arch: string;
    /**
     * Device's OS platform
     */
    platform: 'docker' | 'native';
};

/**
 * Namespace's name
 */
export type NamespaceName = string;

/**
 * Device's status update date
 */
export type DeviceUpdateAt = string;
// ...
export type DeviceTags = Array<Tag>;

/**
 * Device's public URL status.
 */
export type DevicePublicUrl = boolean;
// ...
export type DeviceAcceptable = boolean;

export type Device = {
    uid: DeviceUid;
    /**
     * Device's name
     *
     *
     * By default, the name is the device's MAC address when it just added.
     *
     */
    name: string;
    identity?: DeviceIdentity;
    info?: DeviceInfo;
    /**
     * Device's public key.
     */
    public_key?: string;
    tenant_id: NamespaceTenantId;
    /**
     * Device's last seen date
     */
    last_seen: string;
    /**
     * Device's availability status
     */
    online: boolean;
    namespace?: NamespaceName;
    status: DeviceStatus;
    status_update_at?: DeviceUpdateAt;
    /**
     * Device's creation date
     */
    created_at: string;
    /**
     * Device's remote address
     */
    remote_addr?: string;
    /**
     * Device's geolocation position
     */
    position?: {
        /**
         * Device's latitude position
         */
        latitude?: number;
        /**
         * Device's longitude position
         */
        longitude?: number;
    };
    tags?: DeviceTags;
    public_url?: DevicePublicUrl;
    acceptable?: DeviceAcceptable;
};

/**
 * Device's name
 *
 *
 * By default, the name is the device's MAC address when it just
 * added.
 *
 */
export type DeviceName = string;

/**
 * Session's UID
 */
export type SessionUid = string;

export type Session = {
    uid: SessionUid;
    device_uid?: DeviceUid;
    device: Device | null;
    tenant_id: NamespaceTenantId;
    /**
     * Session's username
     */
    username: string;
    /**
     * Session's IP address
     */
    ip_address: string;
    /**
     * Session's started date
     */
    started_at: string;
    /**
     * Session's last seen date
     */
    last_seen: string;
    /**
     * Session's active status
     */
    active: boolean;
    /**
     * Session's authenticated status
     */
    authenticated: boolean;
    /**
     * Session's recorded status
     */
    recorded: boolean;
    /**
     * Session's type
     */
    type: 'web' | 'term';
    /**
     * Session's terminal
     */
    term: string;
    /**
     * Session's geolocation position
     */
    position: {
        /**
         * Session's latitude position
         */
        latitude?: number;
        /**
         * Session's longitude position
         */
        longitude?: number;
    };
    /**
     * Session's events
     */
    events: {
        /**
         * Session's set of types
         */
        types?: Array<string>;
        /**
         * Session's seats
         */
        seats?: Array<number>;
        /**
         * Session's list of events
         */
        items?: Array<{
            /**
             * The type of the event
             */
            type?: string;
            /**
             * The time the event occurred in ISO 8601 format
             */
            timestamp?: string;
            /**
             * Additional data related to the event
             */
            data?: {
                [key: string]: unknown;
            };
            /**
             * Seat where the event happened
             */
            seat?: number;
        }>;
    };
};
// ...
export type GetSessionsAdminData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * Page number
         */
        page?: number;
        /**
         * Items per page
         */
        per_page?: number;
    };
    url: '/admin/api/sessions';
};